% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{fastqc}
\alias{fastqc}
\title{Run fastqc}
\usage{
fastqc(filepaths, outputDir = "fastqc_output", cmd = "fastqc", args = NULL)
}
\arguments{
\item{filepaths}{Paths to fastq files.}

\item{outputDir}{Directory in which to store fastqc's
output.}

\item{cmd}{Name of fastqc command-line interface.}

\item{args}{Arguments to pass to fastqc's CLI.}
}
\value{
A vector of exit codes, invisibly.
}
\description{
This function calls
\href{https://www.bioinformatics.babraham.ac.uk/projects/fastqc/}{fastqc} using
\code{\link[=system2]{system2()}}. To run in parallel, register a parallel backend using
\code{\link[doFuture:registerDoFuture]{doFuture::registerDoFuture()}} or \code{\link[doParallel:registerDoParallel]{doParallel::registerDoParallel()}}.
}
