% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetchMetadata}
\alias{fetchMetadata}
\title{Fetch metadata for a genomic study}
\usage{
fetchMetadata(
  bioproject,
  host = c("ena", "sra"),
  fields = c("study_accession", "sample_accession", "secondary_sample_accession",
    "sample_alias", "sample_title", "experiment_accession", "run_accession", "fastq_md5",
    "fastq_ftp", "fastq_aspera"),
  file = NULL
)
}
\arguments{
\item{bioproject}{String indicating bioproject accession.}

\item{host}{String indicating from where to fetch the metadata.}

\item{fields}{Character vector indicating which fields to fetch, if \code{host}
is "ena".}

\item{file}{String indicating output file path, if not \code{NULL}.}
}
\value{
A \code{data.table}.
}
\description{
This function can use the API of the European Nucleotide Archive
(recommended) or the Sequence Read Archive.
}
\seealso{
\code{\link[=seeker]{seeker()}}, \code{\link[=fetch]{fetch()}}
}
