% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAscpArgs}
\alias{getAscpArgs}
\title{Get ascp arguments}
\usage{
getAscpArgs()
}
\value{
A character vector.
}
\description{
This function returns the default arguments to pass to the aspera ascp
command-line interface, based on the operating system. Windows is not
supported.
}
\seealso{
\code{\link[=getAscpCmd]{getAscpCmd()}}, \code{\link[=fetch]{fetch()}}
}
