% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAsperaArgs}
\alias{getAsperaArgs}
\title{Get aspera arguments}
\usage{
getAsperaArgs()
}
\value{
A character vector.
}
\description{
This function returns the default arguments to pass to the aspera
command-line interface, based on the operating system. Windows is not
supported.
}
\seealso{
\code{\link[=getAsperaCmd]{getAsperaCmd()}}, \code{\link[=getFastq]{getFastq()}}
}
