% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAsperaCmd}
\alias{getAsperaCmd}
\title{Get aspera command}
\usage{
getAsperaCmd()
}
\value{
A string.
}
\description{
This function returns the default path to the aspera ascp command-line
interface, based on the operating system. Windows is not supported.
}
\seealso{
\code{\link[=getAsperaArgs]{getAsperaArgs()}}, \code{\link[=fetch]{fetch()}}
}
