% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{getFastq}
\alias{getFastq}
\title{Fetch fastq files}
\usage{
getFastq(
  remoteFilepaths,
  outputDir = "fastq",
  overwrite = FALSE,
  ftpCmd = "wget",
  ftpArgs = "-q",
  asperaCmd = "~/.aspera/connect/bin/ascp",
  asperaArgs = c("-QT -l 300m -P33001", "-i",
    "~/.aspera/connect/etc/asperaweb_id_dsa.openssh"),
  asperaPrefix = "era-fasp"
)
}
\arguments{
\item{remoteFilepaths}{Character vector of remote filepaths. For single-end
reads, each element of the vector will be a single filepath. For paired-end
reads, each element can be two filepaths separated by ";". If a remote
filepath starts with "fasp", the file will be downloaded using aspera,
otherwise the file will be downloaded using ftp.}

\item{outputDir}{String indicating the local directory in which to save the
files. Will be created if it doesn't already exist.}

\item{overwrite}{Logical indicating whether to overwrite files that already
exist in \code{outputDir}.}

\item{ftpCmd}{String indicating system command for fetching files by ftp.}

\item{ftpArgs}{Character vector indicating arguments to pass to \code{ftpCmd}.}

\item{asperaCmd}{String indicating path to the aspera ascp program.}

\item{asperaArgs}{Character vector indicating arguments to pass to
\code{asperaCmd}.}

\item{asperaPrefix}{String indicating prefix for downloading files by aspera,
i.e., \code{asperaPrefix@remoteFilepath}.}
}
\value{
A \code{data.table}. As the function runs, it updates a tab-delimited log
file in \code{outputDir} called "progress.tsv".
}
\description{
This function can download fastq files using aspera (recommended) or ftp, by
calling command-line interfaces using \code{\link[=system2]{system2()}}. To download files in
parallel, register a parallel backend using \code{\link[doFuture:registerDoFuture]{doFuture::registerDoFuture()}} or
\code{\link[doParallel:registerDoParallel]{doParallel::registerDoParallel()}}.
}
\seealso{
\code{\link[=getMetadata]{getMetadata()}}
}
