% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{getMetadata}
\alias{getMetadata}
\title{Fetch metadata for a genomic study}
\usage{
getMetadata(
  study,
  host = c("ena", "sra"),
  fields = c("study_accession", "sample_accession", "secondary_sample_accession",
    "sample_alias", "sample_title", "experiment_accession", "run_accession", "fastq_ftp",
    "fastq_aspera")
)
}
\arguments{
\item{study}{String indicating study accession.}

\item{host}{String indicating from where to fetch the metadata.}

\item{fields}{Character vector indicating which fields to fetch.}
}
\value{
A \code{data.table}.
}
\description{
This function can use the API of the European Nucleotide Archive
(recommended) or the Sequence Read Archive.
}
\seealso{
\code{\link[=fetch]{fetch()}}
}
