% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_utils.R
\name{getPlatforms}
\alias{getPlatforms}
\title{Get supported microarray platforms}
\usage{
getPlatforms(type = c("cdf", "mapping"))
}
\arguments{
\item{type}{String indicating whether to get supported platforms for
processing raw Affymetrix data using custom CDF or for mapping already
processed data from probes to genes.}
}
\value{
A \code{data.table}.
}
\description{
Get supported microarray platforms
}
