% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon.R
\name{getSalmonMetadata}
\alias{getSalmonMetadata}
\title{Aggregrate metadata from salmon quantifications}
\usage{
getSalmonMetadata(
  outputDir = "salmon_output",
  outputFilename = "meta_info.csv"
)
}
\arguments{
\item{outputDir}{Directory that contains output from salmon.}

\item{outputFilename}{Name of file, which will be saved in \code{outputDir}. If
\code{NULL}, no file is saved.}
}
\value{
A data.table, invisibly.
}
\description{
Aggregrate metadata from salmon quantifications
}
