% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{installTools}
\alias{installTools}
\title{Install seeker system dependencies for Mac and Ubuntu.}
\usage{
installTools(
  sraToolkitPath = "~",
  sraAddToPath = TRUE,
  minicondaPath = "~",
  minicondaEnv = "seeker",
  setSeekerOption = TRUE,
  refgenieDir = "~/genomes",
  salmonIndexes = NULL,
  fastqDir = NULL
)
}
\arguments{
\item{sraToolkitPath}{Where to install the SRA Toolkit.}

\item{sraAddToPath}{Boolean to add the SRA toolkit to the system and R path files.}

\item{minicondaPath}{Where to install miniconda.}

\item{minicondaEnv}{The name of the miniconda environment. "base" will use the default base environment.}

\item{setSeekerOption}{Boolean to add the miniconda path as the "seeker.miniconda" option to the R profile.}

\item{refgenieDir}{Where to add refgenie config files.}

\item{salmonIndexes}{Which salmon indexes to pull from refgenie.}

\item{fastqDir}{Where to output the fastq_screen command.}
}
\value{
\code{invisible()}
}
\description{
Install seeker system dependencies for Mac and Ubuntu.
}
