% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{multiqc}
\alias{multiqc}
\title{Run MultiQC}
\usage{
multiqc(
  parentDir = ".",
  outputDir = "multiqc_output",
  cmd = "multiqc",
  args = NULL
)
}
\arguments{
\item{parentDir}{Directory that contains output to be aggregated.}

\item{outputDir}{Directory in which to store output. Will be created if it
doesn't exist.}

\item{cmd}{Name or path of the command-line interface.}

\item{args}{Additional arguments to pass to the command-line interface.}
}
\value{
An exit code, invisibly.
}
\description{
This function calls \href{https://multiqc.info/}{multiqc} using \code{\link[=system2]{system2()}}.
}
\seealso{
\code{\link[=seeker]{seeker()}}
}
