% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{seekerArray}
\alias{seekerArray}
\title{Process microarray data end to end}
\usage{
seekerArray(params, parentDir)
}
\arguments{
\item{params}{Named list of parameters with components:
\itemize{
\item \code{study}: String indicating the study accession and used to name the output
directory within \code{parentDir}. If \code{study} starts with 'GSE', data will be
fetched using \code{\link[GEOquery:getGEO]{GEOquery::getGEO()}}. Otherwise the data will be fetched
using \code{\link[ArrayExpress:getAE]{ArrayExpress::getAE()}}.
\item \code{geneIdType}: String indicating whether to map probes to Ensembl Gene IDs
('ensembl') or Entrez Gene IDs ('entrez').
}

\code{params} can be derived from a yaml file, see
\code{vignette('introduction', package = 'seeker')}. The yaml representation
of \code{params} will be saved to \code{parentDir}/\code{params$study}/params.yml.}

\item{parentDir}{Directory in which to store the output, which will be a
directory named according to \code{params$study}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
This function fetches data and metadata from NCBI GEO and ArrayExpress,
processes raw Affymetrix data using RMA and custom CDFs from Brainarray, and
maps probes to genes.
}
