% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkDefaultCommands}
\alias{checkDefaultCommands}
\title{Check for presence of command-line interfaces}
\usage{
checkDefaultCommands(keepIdx = FALSE)
}
\arguments{
\item{keepIdx}{Logical indicating whether to keep the \code{idx} column of the
resulting data.table. For internal use only.}
}
\value{
A data.table with columns for command, path, and version.
}
\description{
This function checks whether the command-line tools used by seeker are
accessible in the expected places.
}
\seealso{
\code{\link[=installSysDeps]{installSysDeps()}}
}
