% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon.R
\name{getSalmonMetadata}
\alias{getSalmonMetadata}
\title{Aggregrate metadata from salmon quantifications}
\usage{
getSalmonMetadata(inputDir, outputDir = "data")
}
\arguments{
\item{inputDir}{Directory that contains output from salmon.}

\item{outputDir}{Directory in which to save the result, a file named
"salmon_meta_info.csv". If \code{NULL}, no file is saved.}
}
\value{
A data.table, invisibly.

#' @seealso \code{\link[=seeker]{seeker()}}, \code{\link[=salmon]{salmon()}}
}
\description{
Aggregrate metadata from salmon quantifications
}
