% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon.R
\name{getSalmonMetadata}
\alias{getSalmonMetadata}
\title{Aggregrate metadata from salmon quantifications}
\usage{
getSalmonMetadata(
  outputDir = "salmon_output",
  outputFilepath = "salmon_meta_info.csv"
)
}
\arguments{
\item{outputDir}{Directory that contains output from salmon.}

\item{outputFilepath}{Path to which to save the result. If \code{NULL}, no file is
saved.}
}
\value{
A data.table, invisibly.
}
\description{
Aggregrate metadata from salmon quantifications
}
