% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tximport.R
\name{getTx2gene}
\alias{getTx2gene}
\title{Get mapping between transcripts and genes}
\usage{
getTx2gene(
  dataset = "hsapiens_gene_ensembl",
  outputFilepath = "tx2gene.csv",
  ...
)
}
\arguments{
\item{dataset}{Ensembl gene dataset, passed to \code{\link[biomaRt:useEnsembl]{biomaRt::useEnsembl()}}. To
see which datasets are available, do
\verb{mart = useEnsembl('genes'); listDatasets(mart)}.}

\item{outputFilepath}{Path to which to save the result. If \code{NULL}, no file is
saved.}

\item{...}{Additional arguments passed to \code{\link[biomaRt:useEnsembl]{biomaRt::useEnsembl()}}.}
}
\value{
A data.frame, as returned by \code{\link[biomaRt:getBM]{biomaRt::getBM()}}.
}
\description{
This function uses the
\href{https://doi.org/doi:10.18129/B9.bioc.biomaRt}{biomaRt package}.
}
