% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tximport.R
\name{getTx2gene}
\alias{getTx2gene}
\title{Get mapping between transcripts and genes}
\usage{
getTx2gene(
  dataset = "mmusculus_gene_ensembl",
  version = NULL,
  outputDir = "data"
)
}
\arguments{
\item{dataset}{Ensembl gene dataset, passed to \code{\link[biomaRt:useEnsembl]{biomaRt::useEnsembl()}}. To
see which datasets are available, do
\verb{mart = biomaRt::useEnsembl('genes'); biomaRt::listDatasets(mart)}.}

\item{version}{Passed to \code{\link[biomaRt:useEnsembl]{biomaRt::useEnsembl()}}.}

\item{outputDir}{Directory in which to save the result, a file named
"tx2gene.csv.gz". If \code{NULL}, no file is saved.}
}
\value{
A data.frame, as returned by \code{\link[biomaRt:getBM]{biomaRt::getBM()}}.
}
\description{
This function uses the
\href{https://doi.org/doi:10.18129/B9.bioc.biomaRt}{biomaRt package}.
}
