% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_sys_deps.R
\name{installSysDeps}
\alias{installSysDeps}
\title{Install seeker system dependencies for Mac and Ubuntu.}
\usage{
installSysDeps(
  sraToolkitDir = "~",
  minicondaDir = "~",
  minicondaEnv = "seeker",
  refgenieDir = "~/refgenie_genomes",
  salmonIndexes = NULL,
  fastqscreenDir = NULL
)
}
\arguments{
\item{sraToolkitDir}{Where to install the SRA Toolkit.}

\item{minicondaDir}{Where to install miniconda.}

\item{minicondaEnv}{The name of the miniconda environment. "base" will use the default base environment.}

\item{refgenieDir}{Where to add refgenie config files.}

\item{salmonIndexes}{Which salmon indexes to pull from refgenie.}

\item{fastqscreenDir}{Where to output the fastq_screen command.}
}
\value{
\code{invisible()}
}
\description{
Install seeker system dependencies for Mac and Ubuntu.
}
