% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_sys_deps.R
\name{installSysDeps}
\alias{installSysDeps}
\title{Install seeker's system dependencies}
\usage{
installSysDeps(
  sraToolkitDir = "~",
  minicondaDir = "~",
  minicondaEnv = "seeker",
  refgenieDir = "~/refgenie_genomes",
  refgenieGenomes = NULL,
  fastqscreenDir = NULL
)
}
\arguments{
\item{sraToolkitDir}{String indicating directory in which to install the
\href{https://github.com/ncbi/sra-tools}{SRA Toolkit}. If \code{NULL}, the Toolkit
will not be installed.}

\item{minicondaDir}{String indicating directory in which to install
\href{https://docs.conda.io/en/latest/miniconda.html}{Miniconda}. If \code{NULL},
Miniconda will not be installed.}

\item{minicondaEnv}{String indicating name of the Miniconda environment in
which to install various conda packages (fastq-screen, fastqc, multiqc,
pigz, refgenie, salmon, and trim-galore).}

\item{refgenieDir}{String indicating directory in which to download genome
assets using refgenie. Only used if \code{minicondaDir} is not \code{NULL}.}

\item{refgenieGenomes}{Character vector indicating genome assets, such as
transcriptome indexes for \code{\link[=salmon]{salmon()}}, to pull from
\href{http://refgenomes.databio.org/index}{refgenomes} using refgenie. If
\code{NULL}, no assets are fetched.}

\item{fastqscreenDir}{String indicating directory in which to download the
genomes for \code{\link[=fastqscreen]{fastqscreen()}}. This takes a long time. If \code{NULL}, genomes are
not downloaded.}
}
\value{
\code{NULL}, invisibly
}
\description{
This function installs and configures the various programs required for
seeker to fetch and process RNA-seq data.
}
\seealso{
\code{\link[=seeker]{seeker()}}
}
