% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeker.R
\name{seeker}
\alias{seeker}
\title{Process RNA-seq data end to end}
\usage{
seeker(params, parentDir = ".")
}
\arguments{
\item{params}{Named list of parameters with components:
\itemize{
\item \code{study}: String used to name the output directory within \code{parentDir}.
\item \code{metadata}: Named list with components:
\itemize{
\item \code{run}: Logical indicating whether to fetch metadata from ENA. See
\code{\link[=fetchMetadata]{fetchMetadata()}}. If \code{TRUE}, saves a file
\code{parentDir}/\code{study}/data/metadata.csv. If \code{FALSE}, expects that file to
already exist. Following components are only checked if \code{run} is \code{TRUE}.
\item \code{bioproject}: String indicating the study's bioproject accession.
\item \code{include}: Optional named list for specifying which rows of metadata to
include for further processing, with components:
\itemize{
\item \code{colname}: String indicating column in metadata
\item \code{values}: Vector indicating values within \code{colname}
}
\item \code{exclude}: Optional named list for specifying which rows of metadata to
exclude from further processing (superseding \code{include}), with components:
\itemize{
\item \code{colname}: String indicating column in metadata
\item \code{values}: Vector indicating values within \code{colname}
}
}
\item \code{fetch}: Named list with components:
\itemize{
\item \code{run}: Logical indicating whether to fetch fastq(.gz) files using ascp.
See \code{\link[=fetch]{fetch()}}. If \code{TRUE}, expects metadata to have a column
'fastq_aspera' containing remote paths, and saves files to
\code{parentDir}/\code{study}/fetch_output. If \code{FALSE}, expects metadata to have a
column 'fastq_aspera' containing names (or complete paths, local or
remote) of fastq files. Whether \code{TRUE} or \code{FALSE}, updates metadata with
column 'fastq_fetched' containing paths to files that should be in
\code{parentDir}/\code{study}/fetch_output. Following components are only checked
if \code{run} is \code{TRUE}.
\item \code{overwrite}: Logical indicating whether to overwrite files that already
exist. \code{NULL} indicates to use the default in \code{\link[=fetch]{fetch()}}.
\item \code{ascpCmd}: String indicating path to ascp. \code{NULL} indicates to use the
default in \code{\link[=fetch]{fetch()}}.
\item \code{ascpArgs}: Character vector of arguments to pass to ascp. \code{NULL}
indicates to use the default in \code{\link[=fetch]{fetch()}}.
\item \code{ascpPrefix}: String indicating prefix for fetching files. \code{NULL}
indicates to use the default in \code{\link[=fetch]{fetch()}}.
}
\item \code{trimgalore}: Named list with components:
\itemize{
\item \code{run}: Logical indicating whether to perform quality/adapter trimming of
reads. See \code{\link[=trimgalore]{trimgalore()}}. If \code{TRUE}, expects metadata to have a column
'fastq_fetched' containing paths to fastq files in
\code{parentDir}/\code{study}/fetch_output, saves trimmed files to
\code{parentDir}/\code{study}/trimgalore_output, and updates metadata with column
'fastq_trimmed'. If \code{FALSE}, expects and does nothing. Following
components are only checked if \code{run} is \code{TRUE}.
\item \code{cmd}: Name or path of the command-line interface. \code{NULL} indicates to
use the default in \code{\link[=trimgalore]{trimgalore()}}.
\item \code{args}: Additional arguments to pass to the command-line interface.
\code{NULL} indicates to use the default in \code{\link[=trimgalore]{trimgalore()}}.
}
\item \code{fastqc}: Named list with components:
\itemize{
\item \code{run}: Logical indicating whether to perform QC on reads. See \code{\link[=fastqc]{fastqc()}}.
If \code{TRUE} and \code{trimgalore$run} is \code{TRUE}, expects metadata to have a
column 'fastq_trimmed' containing paths to fastq files in
\code{parentDir}/\code{study}/trimgalore_output. If \code{TRUE} and \code{trimgalore$run} is
\code{FALSE}, expects metadata to have a column 'fastq_fetched' containing
paths to fastq files in \code{parentDir}/\code{study}/fetch_output. If \code{TRUE},
saves results to \code{parentDir}/\code{study}/fastqc_output. If \code{FALSE}, expects
and does nothing. Following components are only checked if \code{run} is
\code{TRUE}.
\item \code{cmd}: Name or path of the command-line interface. \code{NULL} indicates to
use the default in \code{\link[=fastqc]{fastqc()}}.
\item \code{args}: Additional arguments to pass to the command-line interface.
\code{NULL} indicates to use the default in \code{\link[=fastqc]{fastqc()}}.
}
\item \code{salmon}: Named list with components:
\itemize{
\item \code{run}: Logical indicating whether to quantify transcript abundances. See
\code{\link[=salmon]{salmon()}}. If \code{TRUE} and \code{trimgalore$run} is \code{TRUE}, expects metadata to
have a column 'fastq_trimmed' containing paths to fastq files in
\code{parentDir}/\code{study}/trimgalore_output. If \code{TRUE} and \code{trimgalore$run} is
\code{FALSE}, expects metadata to have a column 'fastq_fetched' containing
paths to fastq files in \code{parentDir}/\code{study}/fetch_output. If \code{TRUE}, also
expects metadata to have a column 'sample_accession' containing sample
ids, and saves results to \code{parentDir}/\code{study}/salmon_output and
\code{parentDir}/\code{study}/data/salmon_meta_info.csv. If \code{FALSE}, expects and
does nothing. Following components are only checked if \code{run} is \code{TRUE}.
\item \code{indexDir}: Directory that contains salmon index.
\item \code{cmd}: Name or path of the command-line interface. \code{NULL} indicates to
use the default in \code{\link[=salmon]{salmon()}}.
\item \code{args}: Additional arguments to pass to the command-line interface.
\code{NULL} indicates to use the default in \code{\link[=salmon]{salmon()}}.
}
\item \code{multiqc}: Named list with components:
\itemize{
\item \code{run}: Logical indicating whether to aggregrate results of various
processing steps. See \code{\link[=multiqc]{multiqc()}}. If \code{TRUE}, saves results to
\code{parentDir}/\code{study}/multiqc_output. If \code{FALSE}, expects and does nothing.
Following components are only checked if \code{run} is \code{TRUE}.
\item \code{cmd}: Name or path of the command-line interface. \code{NULL} indicates to
use the default in \code{\link[=multiqc]{multiqc()}}.
\item \code{args}: Additional arguments to pass to the command-line interface.
\code{NULL} indicates to use the default in \code{\link[=multiqc]{multiqc()}}.
}
\item \code{tximport}: Named list with components:
\itemize{
\item \code{run}: Logical indicating whether to summarize transcript- or gene-level
estimates for downstream analysis. See \code{\link[=tximport]{tximport()}}. If \code{TRUE}, expects
metadata to have a column 'sample_accession' of sample ids, and expects a
directory \code{parentDir}/\code{study}/salmon_output containing directories of
quantification results, and saves results to
\code{parentDir}/\code{study}/data/tximport_output.qs. If \code{FALSE}, expects and does
nothing. Following components are only checked if \code{run} is \code{TRUE}.
\item \code{tx2gene}: Optional named list with components:
\itemize{
\item \code{dataset}: String indicating ensembl gene dataset. See \code{\link[=getTx2gene]{getTx2gene()}}.
\item \code{version}: Number indicating ensembl version. See \code{\link[=getTx2gene]{getTx2gene()}}.
}

If specified, saves a file \code{parentDir}/\code{study}/data/tx2gene.csv.
\item \code{countsFromAbundance}: String indicating whether or how to estimate
counts using estimated abundances. See \code{\link[tximport:tximport]{tximport::tximport()}}.
\item \code{ignoreTxVersion}: Logical indicating whether to the version suffix on
transcript ids. If \code{NULL}, indicates to use \code{TRUE}. See
\code{\link[tximport:tximport]{tximport::tximport()}}.
}
}

\code{params} can be derived from a yaml file, see
\code{vignette('introduction', package = 'seeker')}. The yaml representation
of \code{params} will be saved to \code{parentDir}/\code{params$study}/data/params.yml.}

\item{parentDir}{Directory in which to store the output, which will be a
directory named according to \code{params$study}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
This function selectively performs various steps to process RNA-seq data.
}
\seealso{
\code{\link[=fetchMetadata]{fetchMetadata()}}, \code{\link[=fetch]{fetch()}}, \code{\link[=trimgalore]{trimgalore()}}, \code{\link[=fastqc]{fastqc()}},
\code{\link[=salmon]{salmon()}}, \code{\link[=multiqc]{multiqc()}}, \code{\link[=tximport]{tximport()}}
}
