% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tximport.R
\name{tximport}
\alias{tximport}
\title{Run tximport on RNA-seq quantifications}
\usage{
tximport(
  dirpaths,
  tx2gene,
  outputFilepath = "tximport_output.qs",
  type = c("salmon", "kallisto"),
  countsFromAbundance = "lengthScaledTPM",
  ignoreTxVersion = TRUE,
  ...
)
}
\arguments{
\item{dirpaths}{Paths to directories than contain quantification results.}

\item{tx2gene}{Data.frame, as returned by \code{\link[=getTx2gene]{getTx2gene()}}, passed to
\code{\link[tximport:tximport]{tximport::tximport()}}.}

\item{outputFilepath}{Path to which to save the result using \code{\link[qs:qsave]{qs::qsave()}}.
If \code{NULL}, no file will be saved.}

\item{type}{Passed to \code{\link[tximport:tximport]{tximport::tximport()}}.}

\item{countsFromAbundance}{Passed to \code{\link[tximport:tximport]{tximport::tximport()}}.}

\item{ignoreTxVersion}{Passed to \code{\link[tximport:tximport]{tximport::tximport()}}.}

\item{...}{Additional arguments passed to \code{\link[tximport:tximport]{tximport::tximport()}}.}
}
\value{
A list, as returned by \code{\link[tximport:tximport]{tximport::tximport()}}, invisibly.
}
\description{
This function uses the
\href{https://doi.org/doi:10.18129/B9.bioc.tximport}{tximport package}.
}
